fix logger circular dependency
