+++ libraries/aleth/libp2p/Host.cpp
@@ -12,6 +12,7 @@
 
 #include "Capability.h"
 #include "Common.h"
+#include "LoggerFormatters.hpp"
 #include "RLPxHandshake.h"
 #include "Session.h"
