+++ libraries/core_libs/consensus/src/final_chain/final_chain.cpp
@@ -5,6 +5,7 @@
 #include <utility>
 
 #include "common/encoding_solidity.hpp"
+#include "common/logger_formatters.hpp"
 #include "common/types.hpp"
 #include "common/util.hpp"
 #include "final_chain/state_api_data.hpp"
