+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/latest/common/base_packet_handler.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "common/logger_formatters.hpp"
 #include "network/threadpool/packet_data.hpp"
 
 namespace taraxa::network::tarcap {
