+++ libraries/logger/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(HEADERS
     include/logger/logging.hpp
+    include/logger/logging_config.hpp
 )
 
 set(SOURCES
@@ -9,7 +10,7 @@ set(SOURCES
 find_package(spdlog REQUIRED)
 add_library(logger ${SOURCES} ${HEADERS})
 target_include_directories(logger PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
-target_link_libraries(logger PUBLIC spdlog::spdlog config)
+target_link_libraries(logger PUBLIC spdlog::spdlog)
 
 install(TARGETS logger
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
